adding rest for multiple invoices

+++ src/main/java/com/endava/service_system/service/InvoiceService.java
@@ -188,7 +188,7 @@ public class InvoiceService {
         makeInvoicesPayed(ids);
     }
 
-    public Optional<InvoiceForPaymentDto> getFullInvoiceById(int id) {
+    public Optional<InvoiceForPaymentDto> getFullInvoiceById(Long id) {
         return invoiceDao.getFullInvoiceById(id);
     }
